requirements.txt
asgiref==3.7.2
Django==5.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.3.0

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
django-redis==5.4.0
dj-database-url==2.1.0